// src/interfaces/sect.ts
// 流派列表
export interface Sect {
	'火': '燃烧' | '火弹' | '火环' | '地雷' | '火精灵';
	'冰': '寒冷' | '寒冷(爆发)' | '寒冷(冰冻)' | '冰锥' | '冰刺';
	'电': '感电' | '闪电链' | '落雷' | '电球' | '雷云';
	'毒': '中毒' | '中毒(毒雾)' | '中毒(毒环)' | '毒弹' | '毒液' | '史莱姆';
	'暗': '触手' | '影子' | '影刺' | '黑洞';
	'光': '光枪' | '闪光' | '光波' | '光阵' | '圣光标记';
	'刃': '飞剑' | '撕裂' | '刃环' | '刀刃风暴';
}